doit: Split out cmd_debug
